CREATE: Create Exceptions

+++ app/server/database.py
@@ -1,14 +1,17 @@
 from motor import motor_asyncio
 from .models import Diary
+from app.server.exceptions import *
 
 host = 'localhost'
 port = 27017
 
-client = motor_asyncio.AsyncIOMotorClient(host,port)
+#database settings
+def get_db():
+    client = motor_asyncio.AsyncIOMotorClient(host,port)
+    db = client.iaryda
+    return db
 
-db = client.iaryda
-
-diary_collection = db.get_collection("diary")
+diary_collection = get_db().get_collection("diary")
 
 #helpers
 def diary_helper(diary) -> dict:
@@ -19,6 +22,12 @@ def feeling_helper(diary) -> dict:
     diary.pop("_id")
     diary.pop("content")
     return diary
+
+def check_feeling_validation(diary):
+    correct = ["very_bad", "bad", "so_so","good","very_good"]
+    if not diary["feeling"] in correct:
+        return False
+    return True
 
 #db 조회, 관리 함수 만들어야함
 async def get_diaries() -> list:
@@ -32,29 +41,50 @@ async def get_diaries() -> list:
 
 async def get_diary(date: str) -> dict:
     diary = await diary_collection.find_one({"date" : date})
-    return diary_helper(diary)
-
+    if diary != None:
+        return diary_helper(diary)
+    else:
+        raise DiaryDoseNotExistExecption()
 
 async def create_diary(diary_data: dict) -> dict:
-    diary = await diary_collection.insert_one(diary_data)
-    result = await get_diary(diary_data['date'])
-    return result
+    if not await diary_collection.find_one({"date" : diary_data["date"]}) == None:
+        raise DiaryAlreadyExistException()
+    elif not check_feeling_validation(diary_data):
+        raise UnvalidFeelingException()
+    else:
+        diary = await diary_collection.insert_one(diary_data)
+        result = await get_diary(diary_data['date'])
+        return result
+    
 
 async def update_diary(date:str , diary_data: dict) -> dict:
-    diary = await diary_collection.update_one({"date" : date}, {"$set" : diary_data})
-    result = await get_diary(date)
-    return result
+    if await diary_collection.find_one({"date" : date}) == None:
+        raise DiaryDoseNotExistExecption()
+    elif not check_feeling_validation(diary_data):
+        raise UnvalidFeelingException()
+    else:
+        diary = await diary_collection.update_one({"date" : date}, {"$set" : diary_data})
+        result = await get_diary(date)
+        return result
+    
+
 
 async def delete_diary(date:str):
-    await diary_collection.delete_one({"date" : date})
-    return True
+    if await diary_collection.find_one({"date" : date}) == None:
+        raise DiaryDoseNotExistExecption()
+    else:
+        await diary_collection.delete_one({"date" : date})
+        return []
 
 async def search_diary(query):
-    result_data = []
-        
-    async for diary in diary_collection.find(query):
-        result_data.append(diary_helper(diary))
-    return result_data
+    if "feeling" in query and not check_feeling_validation(query):
+        raise UnvalidFeelingException()
+    else:
+        result_data = []
+
+        async for diary in diary_collection.find(query):
+            result_data.append(diary_helper(diary))
+        return result_data
 
 async def get_feelings(month: str):
     result_data = []
